+++ src/main/java/com/ifgoiano/urt/projetocaoguia/projetocaoguiabackend/formularios/model/FormularioRequestDTO.java
@@ -1,4 +1,10 @@
 package com.ifgoiano.urt.projetocaoguia.projetocaoguiabackend.formularios.model;
 
-public record FormularioRequestDTO(String resposta) {
+public record FormularioRequestDTO(
+        String nome,
+        String email,
+        String telefone,
+        CategoriaFormulario categoria,
+        String resposta
+) {
 }