+++ src/main/java/com/ifgoiano/urt/projetocaoguia/projetocaoguiabackend/usuarios/model/Usuario.java
@@ -22,6 +22,9 @@ public class Usuario {
     @Column(nullable = false, unique = true)
     private String email;
 
+    @Column(length = 20)
+    private String telefone;
+
     @Column(nullable = false)
     private String senha;